submit

+++ scripts/deploy-spark.sh
@@ -125,10 +125,14 @@ fi
 #
 # spark on hive 配置
 #
+if [ ! -f "$SPARK_HOME/conf/hive-site.xml" ]; then
+    log_info "配置 Spark on hive"
+    cp $HIVE_HOME/conf/hive-site.xml $SPARK_HOME/conf/
+    cp $HIVE_HOME/lib/mysql-connector-java-5.1.27-bin.jar $SPARK_HOME/jars/
+    cp $HADOOP_HOME/etc/hadoop/core-site.xml $SPARK_HOME/conf/
+    cp $HADOOP_HOME/etc/hadoop/hdfs-site.xml $SPARK_HOME/conf/
 
-cp $HIVE_HOME/conf/hive-site.xml $SPARK_HOME/conf/
-cp $HIVE_HOME/lib/mysql-connector-java-5.1.27-bin.jar $SPARK_HOME/jars/
-cp $HADOOP_HOME/etc/hadoop/core-site.xml $SPARK_HOME/conf/
-cp $HADOOP_HOME/etc/hadoop/hdfs-site.xml $SPARK_HOME/conf/
+    chown marmot:marmot -R $SPARK_HOME
+fi
 
-chown marmot:marmot -R $SPARK_HOME
+
